chore: move tests to resolvers dir
